+++ src/org/atoum/intellij/plugin/atoum/run/CommandLineArgumentsBuilder.java
@@ -25,6 +25,14 @@ public class CommandLineArgumentsBuilder {
         return this;
     }
 
+    public CommandLineArgumentsBuilder useConfigFile(String configFilePath)
+    {
+        this.commandLineArgs.add("-c");
+        this.commandLineArgs.add(configFilePath);
+
+        return this;
+    }
+
     public CommandLineArgumentsBuilder useConfiguration(RunnerConfiguration runnerConfiguration)
     {
         if (null != runnerConfiguration.getDirectory()) {